Move mock client generator to package
